refactor: refactor getting twap order status
